+++ project/src/components/app/app.tsx
@@ -1,4 +1,10 @@
+import {BrowserRouter, Route, Routes} from 'react-router-dom';
 import Main from '../../pages/main/main';
+import Login from '../../pages/login/login';
+import Room from '../../pages/offer/offer';
+import NotFound from '../not-found/not-found';
+import {AppRoute} from '../../const';
+import { HelmetProvider } from 'react-helmet-async';
 
 type AppProps = {
   offers: number;
@@ -6,7 +12,28 @@ type AppProps = {
 
 function App({offers}: AppProps): JSX.Element {
   return (
-    <Main offers={offers} />
+    <HelmetProvider>
+      <BrowserRouter>
+        <Routes>
+          <Route
+            path={AppRoute.Root}
+            element={<Main offers={offers} />}
+          />
+          <Route
+            path={AppRoute.Login}
+            element={<Login />}
+          />
+          <Route
+            path={AppRoute.Room}
+            element={<Room />}
+          />
+          <Route
+            path='*'
+            element={<NotFound />}
+          />
+        </Routes>
+      </BrowserRouter>
+    </HelmetProvider>
   );
 }
