+++ project/src/const.ts
@@ -1,0 +1,5 @@
+export enum AppRoute {
+  Root = '/',
+  Login = '/login',
+  Room = '/offer/:id'
+}
